Update import.rb
The Kapp Category Attribute Definitions and Categories import was always deleting the values that existed on the destination server but not in the import data.

Added the condition vars["options"]["delete"] && to the qualification of the delete

+++ import.rb
@@ -519,7 +519,7 @@ Dir["#{core_path}/space/kapps/*"].each { |file|
     # Delete Kapp Category Definitions
     # ------------------------------------------------------------------------------
     destinationKappAttributeArray.each { | attribute |
-      if !sourceKappCategoryArray.include?(attribute)
+      if vars["options"]["delete"] && !sourceKappCategoryArray.include?(attribute)
           space_sdk.delete_category_attribute_definition(kapp['slug'],attribute)
       end
     }
@@ -618,7 +618,7 @@ Dir["#{core_path}/space/kapps/*"].each { |file|
     # ------------------------------------------------------------------------------
      
     destinationCategoryArray.each { | attribute |
-      if !sourceCategoryArray.include?(attribute)
+      if vars["options"]["delete"] && !sourceCategoryArray.include?(attribute)
           space_sdk.delete_category_on_kapp(kapp['slug'],attribute)
       end
     }
